refactor(editor): table structure 모델을 분리한다
