+++ front/src/components/editor/BlockEditorEngine.tsx
@@ -114,6 +114,14 @@ import {
   shouldPromoteWideTableOverflowMode,
   shrinkTableColumnWidthsToFit,
 } from "./tableWidthModel"
+import {
+  type TableColumnCellRef,
+  buildReorderedSimpleTableNode,
+  canShrinkTableAxisAtEnd,
+  collectSimpleTableColumnCells,
+  countShrinkableRenderedTableAxisAtEnd,
+  countShrinkableTableAxisAtEnd,
+} from "./tableStructureModel"
 
 type RuntimeGuardWindow = Window & {
   __AQ_RUNTIME_GUARD_ENABLED__?: boolean
@@ -782,168 +790,6 @@ const TABLE_TRAILING_ADD_EDGE_HOTZONE_PX = 18
 const TABLE_OVERFLOW_COACHMARK_ESTIMATED_WIDTH_PX = 244
 const TABLE_OVERFLOW_COACHMARK_DISMISS_MS = 4200
 
-const isTableCellNode = (node: ProseMirrorNode | null | undefined) =>
-  Boolean(node && (node.type.name === "tableCell" || node.type.name === "tableHeader"))
-
-const isVisuallyEmptyTableCellContent = (node: ProseMirrorNode | null | undefined): boolean => {
-  if (!node) return true
-  if (node.isText) return node.textContent.trim().length === 0
-  if (node.isLeaf) return false
-
-  for (let childIndex = 0; childIndex < node.childCount; childIndex += 1) {
-    if (!isVisuallyEmptyTableCellContent(node.child(childIndex))) {
-      return false
-    }
-  }
-
-  return true
-}
-
-const canShrinkTableAxisAtEnd = (tableNode: ProseMirrorNode, axis: "row" | "column"): boolean => {
-  const rowCount = tableNode.childCount
-  if (rowCount === 0) return false
-
-  const firstRow = tableNode.firstChild
-  if (!firstRow) return false
-  const columnCount = firstRow.childCount
-  if (columnCount === 0) return false
-
-  for (let rowIndex = 0; rowIndex < rowCount; rowIndex += 1) {
-    const row = tableNode.child(rowIndex)
-    if (row.childCount !== columnCount) return false
-    for (let columnIndex = 0; columnIndex < row.childCount; columnIndex += 1) {
-      const cell = row.child(columnIndex)
-      if (!isTableCellNode(cell)) return false
-      const colspan = typeof cell.attrs?.colspan === "number" ? cell.attrs.colspan : 1
-      const rowspan = typeof cell.attrs?.rowspan === "number" ? cell.attrs.rowspan : 1
-      if (colspan > 1 || rowspan > 1) return false
-    }
-  }
-
-  if (axis === "row") {
-    if (rowCount <= 1) return false
-    const trailingRow = tableNode.child(rowCount - 1)
-    for (let columnIndex = 0; columnIndex < trailingRow.childCount; columnIndex += 1) {
-      if (!isVisuallyEmptyTableCellContent(trailingRow.child(columnIndex))) {
-        return false
-      }
-    }
-    return true
-  }
-
-  if (columnCount <= 1) return false
-  const trailingColumnIndex = columnCount - 1
-  for (let rowIndex = 0; rowIndex < rowCount; rowIndex += 1) {
-    const row = tableNode.child(rowIndex)
-    if (!isVisuallyEmptyTableCellContent(row.child(trailingColumnIndex))) {
-      return false
-    }
-  }
-  return true
-}
-
-const countShrinkableTableAxisAtEnd = (tableNode: ProseMirrorNode, axis: "row" | "column"): number => {
-  const rowCount = tableNode.childCount
-  if (rowCount === 0) return 0
-
-  const firstRow = tableNode.firstChild
-  if (!firstRow) return 0
-  const columnCount = firstRow.childCount
-  if (columnCount === 0) return 0
-
-  for (let rowIndex = 0; rowIndex < rowCount; rowIndex += 1) {
-    const row = tableNode.child(rowIndex)
-    if (row.childCount !== columnCount) return 0
-    for (let columnIndex = 0; columnIndex < row.childCount; columnIndex += 1) {
-      const cell = row.child(columnIndex)
-      if (!isTableCellNode(cell)) return 0
-      const colspan = typeof cell.attrs?.colspan === "number" ? cell.attrs.colspan : 1
-      const rowspan = typeof cell.attrs?.rowspan === "number" ? cell.attrs.rowspan : 1
-      if (colspan > 1 || rowspan > 1) return 0
-    }
-  }
-
-  if (axis === "row") {
-    let shrinkableRows = 0
-    for (let rowIndex = rowCount - 1; rowIndex >= 1; rowIndex -= 1) {
-      const trailingRow = tableNode.child(rowIndex)
-      let empty = true
-      for (let columnIndex = 0; columnIndex < trailingRow.childCount; columnIndex += 1) {
-        if (!isVisuallyEmptyTableCellContent(trailingRow.child(columnIndex))) {
-          empty = false
-          break
-        }
-      }
-      if (!empty) break
-      shrinkableRows += 1
-    }
-    return shrinkableRows
-  }
-
-  let shrinkableColumns = 0
-  for (let columnIndex = columnCount - 1; columnIndex >= 1; columnIndex -= 1) {
-    let empty = true
-    for (let rowIndex = 0; rowIndex < rowCount; rowIndex += 1) {
-      const row = tableNode.child(rowIndex)
-      if (!isVisuallyEmptyTableCellContent(row.child(columnIndex))) {
-        empty = false
-        break
-      }
-    }
-    if (!empty) break
-    shrinkableColumns += 1
-  }
-  return shrinkableColumns
-}
-
-const isVisuallyEmptyRenderedTableCell = (cell: HTMLTableCellElement | null | undefined) => {
-  if (!cell) return true
-  if (cell.querySelector("img, video, svg, canvas, iframe, object, embed, table")) return false
-  return cell.textContent?.trim().length ? false : true
-}
-
-const countShrinkableRenderedTableAxisAtEnd = (tableElement: HTMLTableElement | null, axis: "row" | "column"): number => {
-  if (!tableElement) return 0
-  const rows = Array.from(tableElement.querySelectorAll("tr")).filter(
-    (row): row is HTMLTableRowElement => row instanceof HTMLTableRowElement
-  )
-  if (rows.length === 0) return 0
-  const columnCount = rows[0]?.cells.length ?? 0
-  if (columnCount === 0) return 0
-
-  for (const row of rows) {
-    if (row.cells.length !== columnCount) return 0
-    for (const cell of Array.from(row.cells)) {
-      const colspan = Number(cell.getAttribute("colspan") || "1")
-      const rowspan = Number(cell.getAttribute("rowspan") || "1")
-      if (colspan > 1 || rowspan > 1) return 0
-    }
-  }
-
-  if (axis === "row") {
-    let shrinkableRows = 0
-    for (let rowIndex = rows.length - 1; rowIndex >= 1; rowIndex -= 1) {
-      const row = rows[rowIndex]
-      if (Array.from(row.cells).some((cell) => !isVisuallyEmptyRenderedTableCell(cell))) break
-      shrinkableRows += 1
-    }
-    return shrinkableRows
-  }
-
-  let shrinkableColumns = 0
-  for (let columnIndex = columnCount - 1; columnIndex >= 1; columnIndex -= 1) {
-    let empty = true
-    for (const row of rows) {
-      if (!isVisuallyEmptyRenderedTableCell(row.cells[columnIndex] ?? null)) {
-        empty = false
-        break
-      }
-    }
-    if (!empty) break
-    shrinkableColumns += 1
-  }
-  return shrinkableColumns
-}
 const TABLE_EDGE_HANDLE_INSET_PX = 6
 const TABLE_CORNER_CLUSTER_GAP_PX = 6
 const TABLE_CORNER_CLUSTER_WIDTH_PX =
@@ -1379,119 +1225,6 @@ const getCurrentEditorReadableWidthPx = (editor?: TiptapEditor | null) => {
 const shouldClampTableWidthBudget = () => {
   if (typeof window === "undefined" || typeof window.matchMedia !== "function") return false
   return !window.matchMedia(DESKTOP_TABLE_RAIL_MEDIA_QUERY).matches
-}
-
-type TableColumnCellRef = {
-  pos: number
-  node: {
-    attrs?: Record<string, unknown>
-  }
-}
-
-const collectSimpleTableColumnCells = (tableNode: { forEach: Function }, tablePos: number) => {
-  const columns: TableColumnCellRef[][] = []
-  let hasMergedCell = false
-
-  tableNode.forEach((rowNode: any, rowOffset: number) => {
-    if (rowNode.type?.name !== "tableRow") return
-    const rowPos = tablePos + 1 + rowOffset
-    let columnIndex = 0
-
-    rowNode.forEach((cellNode: any, cellOffset: number) => {
-      const colspan = Math.max(1, Number(cellNode.attrs?.colspan ?? 1) || 1)
-      const rowspan = Math.max(1, Number(cellNode.attrs?.rowspan ?? 1) || 1)
-      if (colspan !== 1 || rowspan !== 1) {
-        hasMergedCell = true
-        return
-      }
-
-      const cellPos = rowPos + 1 + cellOffset
-      columns[columnIndex] ||= []
-      columns[columnIndex].push({ pos: cellPos, node: cellNode })
-      columnIndex += 1
-    })
-  })
-
-  return hasMergedCell ? null : columns
-}
-
-const moveItemToInsertionIndex = <T,>(items: T[], sourceIndex: number, insertionIndex: number) => {
-  if (sourceIndex < 0 || sourceIndex >= items.length) {
-    return { changed: false, items, nextIndex: sourceIndex }
-  }
-
-  const boundedInsertionIndex = Math.max(0, Math.min(insertionIndex, items.length))
-  if (boundedInsertionIndex === sourceIndex || boundedInsertionIndex === sourceIndex + 1) {
-    return { changed: false, items, nextIndex: sourceIndex }
-  }
-
-  const nextItems = items.slice()
-  const [movedItem] = nextItems.splice(sourceIndex, 1)
-  if (typeof movedItem === "undefined") {
-    return { changed: false, items, nextIndex: sourceIndex }
-  }
-
-  const nextIndex = boundedInsertionIndex > sourceIndex ? boundedInsertionIndex - 1 : boundedInsertionIndex
-  nextItems.splice(nextIndex, 0, movedItem)
-
-  return {
-    changed: true,
-    items: nextItems,
-    nextIndex,
-  }
-}
-
-const buildReorderedSimpleTableNode = (
-  tableNode: ProseMirrorNode,
-  axis: "row" | "column",
-  sourceIndex: number,
-  insertionIndex: number
-) => {
-  const rowNodes: ProseMirrorNode[] = []
-  tableNode.forEach((rowNode: ProseMirrorNode) => {
-    if (rowNode.type.name === "tableRow") {
-      rowNodes.push(rowNode)
-    }
-  })
-
-  if (rowNodes.length === 0) return null
-
-  if (axis === "row") {
-    const movedRows = moveItemToInsertionIndex(rowNodes, sourceIndex, insertionIndex)
-    if (!movedRows.changed) return null
-    return {
-      node: tableNode.type.create(tableNode.attrs, Fragment.fromArray(movedRows.items), tableNode.marks),
-      nextIndex: movedRows.nextIndex,
-    }
-  }
-
-  if (!collectSimpleTableColumnCells(tableNode, 0)) {
-    return null
-  }
-
-  const firstRowCells: ProseMirrorNode[] = []
-  rowNodes[0]?.forEach((cellNode: ProseMirrorNode) => {
-    firstRowCells.push(cellNode)
-  })
-  const movedColumns = moveItemToInsertionIndex(firstRowCells, sourceIndex, insertionIndex)
-  if (!movedColumns.changed) return null
-
-  const nextRows = rowNodes.map((rowNode) => {
-    const cells: ProseMirrorNode[] = []
-    rowNode.forEach((cellNode: ProseMirrorNode) => {
-      cells.push(cellNode)
-    })
-    if (cells.length !== firstRowCells.length) {
-      return rowNode
-    }
-    const movedCells = moveItemToInsertionIndex(cells, sourceIndex, insertionIndex)
-    return rowNode.type.create(rowNode.attrs, Fragment.fromArray(movedCells.items), rowNode.marks)
-  })
-
-  return {
-    node: tableNode.type.create(tableNode.attrs, Fragment.fromArray(nextRows), tableNode.marks),
-    nextIndex: movedColumns.nextIndex,
-  }
 }
 
 const findActiveRenderedTable = (
